save pools query to dedicated file
